+++ integration/server/mocks/use_global_parameters.ts
@@ -32,7 +32,7 @@ const getThemeAllClasses = ({ themes }: Required<ThemeParameter>['theme']) =>
 const getTargetSelector = ({ selector }: Required<ThemeParameter>['theme']) =>
   (Array.isArray(selector) ? selector.join(', ') : selector) ?? 'body';
 
-function setTheme(themeId: string) {
+function applyThemeCSS(themeId: string) {
   const theme = themeParams.themes.find((t) => t.id === themeId);
   const selector = getTargetSelector(themeParams);
   const targets = selector ? document.querySelectorAll<HTMLElement>(selector) : null;
@@ -48,33 +48,25 @@ function setTheme(themeId: string) {
   }
 }
 
-function getBackground(backgroundId?: string) {
-  if (!backgroundId) return '';
-
-  const option = (backgroundParams.options ?? []).find(({ id }) => id === backgroundId);
-
-  return option ? option.background ?? option.color : '';
-}
-
 export function useGlobalsParameters() {
-  const [themeName, setThemeName] = useState<string>(ThemeId.Light);
-  const [backgroundColor, setBackgroundColor] = useState<string | undefined>('white');
+  const [themeId, setThemeId] = useState<string>(ThemeId.Light);
+  const [backgroundId, setBackgroundId] = useState<string | undefined>('white');
 
   /**
    * Handles setting global context values. Stub for theme and background addons
    */
   function setParams(params: URLSearchParams, parameters?: Parameters) {
     const globals = getGlobalParams(params) as Globals;
-    const themeId = globals.theme ?? parameters?.theme?.default ?? themeParams.default ?? ThemeId.Light;
-    const backgroundId = globals.background ?? parameters?.background?.default ?? backgroundParams.default;
-    setThemeName(themeId);
-    setTheme(themeId);
-    setBackgroundColor(getBackground(backgroundId));
+    const backgroundIdFromParams = globals.background ?? parameters?.background?.default ?? backgroundParams.default;
+    setBackgroundId(backgroundIdFromParams);
+    const themeIdFromParams = globals.theme ?? parameters?.theme?.default ?? themeParams.default ?? ThemeId.Light;
+    setThemeId(themeIdFromParams);
+    applyThemeCSS(themeIdFromParams);
   }
 
   return {
-    themeName,
-    backgroundColor,
+    themeId,
+    backgroundId,
     setParams,
   };
 }