+++ packages/charts/src/chart_types/partition_chart/layout/viewmodel/viewmodel.ts
@@ -279,15 +279,6 @@ const rawChildNodes = (
 };
 
 /** @internal */
-export type PanelPlacement = PartitionSmallMultiplesModel;
-
-/**
- * Todo move it to config
- * @internal
- */
-export const panelTitleFontSize = 16;
-
-/** @internal */
 export function shapeViewModel(
   textMeasure: TextMeasure,
   config: Config,
@@ -300,7 +291,7 @@ export function shapeViewModel(
   topGroove: Pixels,
   containerBackgroundColor: Color,
   smallMultiplesStyle: SmallMultiplesStyle,
-  panel: PanelPlacement,
+  panelModel: PartitionSmallMultiplesModel,
 ): ShapeViewModel {
   const {
     width,
@@ -316,7 +307,7 @@ export function shapeViewModel(
     sectorLineWidth,
   } = config;
 
-  const { marginLeftPx, marginTopPx, panelInnerWidth, panelInnerHeight } = panel;
+  const { marginLeftPx, marginTopPx, panel } = panelModel;
 
   const treemapLayout = isTreemap(partitionLayout);
   const mosaicLayout = isMosaic(partitionLayout);
@@ -328,8 +319,8 @@ export function shapeViewModel(
 
   const diskCenter = isSunburst(partitionLayout)
     ? {
-        x: marginLeftPx + panelInnerWidth / 2,
-        y: marginTopPx + panelInnerHeight / 2,
+        x: marginLeftPx + panel.innerWidth / 2,
+        y: marginTopPx + panel.innerHeight / 2,
       }
     : {
         x: marginLeftPx,
@@ -348,8 +339,8 @@ export function shapeViewModel(
     partitionLayout,
     tree,
     topGroove,
-    panelInnerWidth,
-    panelInnerHeight,
+    panel.innerWidth,
+    panel.innerHeight,
     clockwiseSectors,
     specialFirstInnermostSector,
     maxDepth,
@@ -363,8 +354,8 @@ export function shapeViewModel(
 
   // use the smaller of the two sizes, as a circle fits into a square
   const circleMaximumSize = Math.min(
-    panelInnerWidth,
-    panelInnerHeight - (panel.panelTitle.length > 0 ? panelTitleFontSize * 2 : 0),
+    panel.innerWidth,
+    panel.innerHeight - (panel.title.length > 0 ? panel.fontSize * 2 : 0),
   );
   const outerRadius: Radius = Math.min(outerSizeRatio * circleMaximumSize, circleMaximumSize - sectorLineWidth) / 2;
   const innerRadius: Radius = outerRadius - (1 - emptySizeRatio) * outerRadius;
@@ -376,9 +367,9 @@ export function shapeViewModel(
     layers,
     config.sectorLineWidth,
     config.sectorLineStroke,
-    panel.smAccessorValue,
-    panel.index,
-    panel.innerIndex,
+    panelModel.smAccessorValue,
+    panelModel.index,
+    panelModel.innerIndex,
     config.fillLabel,
     sunburstLayout,
     containerBackgroundColor,
@@ -443,8 +434,8 @@ export function shapeViewModel(
         });
   const maxLinkedLabelTextLength = config.linkLabel.maxTextLength;
   const linkLabelViewModels = linkTextLayout(
-    panelInnerWidth,
-    panelInnerHeight,
+    panel.innerWidth,
+    panel.innerHeight,
     textMeasure,
     config,
     nodesWithoutRoom,
@@ -455,8 +446,8 @@ export function shapeViewModel(
     valueFormatter,
     maxLinkedLabelTextLength,
     {
-      x: width * panel.left + panelInnerWidth / 2,
-      y: height * panel.top + panelInnerHeight / 2,
+      x: width * panelModel.left + panel.innerWidth / 2,
+      y: height * panelModel.top + panel.innerHeight / 2,
     },
     containerBackgroundColor,
   );
@@ -481,22 +472,23 @@ export function shapeViewModel(
   // combined viewModel
   return {
     partitionLayout: config?.partitionLayout ?? defaultConfig.partitionLayout,
-    smAccessorValue: panel.smAccessorValue,
-    panelTitle: panel.panelTitle,
-    index: panel.index,
-    innerIndex: panel.innerIndex,
-    width: panel.width,
-    height: panel.height,
-    top: panel.top,
-    left: panel.left,
-    innerRowCount: panel.innerRowCount,
-    innerColumnCount: panel.innerColumnCount,
-    innerRowIndex: panel.innerRowIndex,
-    innerColumnIndex: panel.innerColumnIndex,
-    marginLeftPx: panel.marginLeftPx,
-    marginTopPx: panel.marginTopPx,
-    panelInnerWidth: panel.panelInnerWidth,
-    panelInnerHeight: panel.panelInnerHeight,
+    smAccessorValue: panelModel.smAccessorValue,
+
+    index: panelModel.index,
+    innerIndex: panelModel.innerIndex,
+    width: panelModel.width,
+    height: panelModel.height,
+    top: panelModel.top,
+    left: panelModel.left,
+    innerRowCount: panelModel.innerRowCount,
+    innerColumnCount: panelModel.innerColumnCount,
+    innerRowIndex: panelModel.innerRowIndex,
+    innerColumnIndex: panelModel.innerColumnIndex,
+    marginLeftPx: panelModel.marginLeftPx,
+    marginTopPx: panelModel.marginTopPx,
+    panel: {
+      ...panelModel.panel,
+    },
 
     config,
     layers,