+++ packages/charts/src/utils/themes/dark_theme.ts
@@ -193,4 +193,42 @@ export const DARK_THEME: Theme = {
   background: {
     color: Colors.Transparent.rgba,
   },
+  goal: {
+    minFontSize: 8,
+    maxFontSize: 64,
+    tickLabel: {
+      fontStyle: 'normal',
+      fontFamily: 'sans-serif',
+      fill: 'white',
+    },
+    majorLabel: {
+      fontStyle: 'normal',
+      fontFamily: 'sans-serif',
+      fill: 'white',
+    },
+    minorLabel: {
+      fontStyle: 'normal',
+      fontFamily: 'sans-serif',
+      fill: 'white',
+    },
+    majorCenterLabel: {
+      fontStyle: 'normal',
+      fontFamily: 'sans-serif',
+      fill: 'white',
+    },
+    minorCenterLabel: {
+      fontStyle: 'normal',
+      fontFamily: 'sans-serif',
+      fill: 'white',
+    },
+    targetLine: {
+      stroke: 'white',
+    },
+    tickLine: {
+      stroke: 'snow',
+    },
+    progressLine: {
+      stroke: 'white',
+    },
+  },
 };