+++ storybook/stories/goal/11_gaps.tsx
@@ -10,7 +10,6 @@ import { boolean, number } from '@storybook/addon-knobs';
 import React from 'react';
 
 import { Chart, Goal, Settings } from '@elastic/charts';
-import { config } from '@elastic/charts/src/chart_types/goal_chart/layout/config/config';
 import { BandFillColorAccessorInput } from '@elastic/charts/src/chart_types/goal_chart/specs';
 import { GoalSubtype } from '@elastic/charts/src/chart_types/goal_chart/specs/constants';
 
@@ -50,7 +49,6 @@ export const Example = () => {
         labelMinor="(thousand USD)  "
         centralMajor="280"
         centralMinor={showTarget ? `target: ${target}` : undefined}
-        config={config}
       />
     </Chart>
   );